+++ src/app/core/models/match-table-row.ts
@@ -1,4 +1,5 @@
 export interface MatchTableRow {
+  id: number;
   players: string;
   score: string;
   winner: string;