+++ src/app/shared/components/table/table.component.ts
@@ -16,12 +16,30 @@ export class TableComponent<T> implements OnInit {
 
   ngOnInit() {
     this.initColumns();
+    this.removeColumnId();
   }
 
+  /**
+   * @description Inits columns based on provided data source and shows no data label in header if data source is empty
+   */
   private initColumns(): void {
     this.columns = this.dataSource.length ? Object.keys(this.dataSource[0]) : ['No data'];
   }
 
+  /**
+   * @description Removes ID column from table if table data exists
+   */
+  private removeColumnId(): void {
+    if (this.columns.length && this.columns[0] !== 'No data') {
+      const indexOfId = this.columns.indexOf('id');
+      this.columns.splice(indexOfId, 1);
+    }
+  }
+
+  /**
+   * @param row Row clicked
+   * @description Emits row click output on row click
+   */
   public onRowClick(row: T): void {
     this.rowClick.emit(row);
   }