+++ src/app/shared/components/title-bar/title-bar.component.ts
@@ -15,6 +15,9 @@ export class TitleBarComponent {
 
   constructor() {}
 
+  /**
+   * @description Emits button click output on button click
+   */
   public onButtonClick(): void {
     this.buttonClick.emit();
   }