+++ src/app/shared/pipes/camel-case-split/camel-case-split.pipe.spec.ts
@@ -1,0 +1,9 @@
+import { CamelCaseSplitPipe } from './camel-case-split.pipe';
+
+describe('CamelCaseSplitPipe', () => {
+  const pipe = new CamelCaseSplitPipe();
+
+  it('should transform "abcDef" to "abc Def"', () => {
+    expect(pipe.transform('abcDef')).toBe('abc Def');
+  });
+});